Issue #80: Split up the sample document definition spec
To help with readability and maintenance, each document type in the sample document definitions now has its own separate specification file. Furthermore, a new helper module was introduced that contains functions that are common to multiple sample document definitions' specifications.
